comm064
Correção de bug: não estava sendo deletada a descrição adicional quando se deletava o carregamento, foi isso que o argenta conseguiu destruir o código. Tratamento de exceções nas aba carregamento para carregamentos fora do nó. Compatibilização da classe float para as entradas do usuário no método da rigidez direta

+++ BackEnd/dimensionamento/core/Auxiliar/Rigidez_Direita.py
@@ -20,7 +20,7 @@ def entrada_test(data: dict, apoio: dict, comprimentoTotal:float):
         #Definicao elementos pelo carregamentos
         for chave in carregamento.keys():
             temp = carregamento[chave]['pos']
-            if isinstance(temp,int):
+            if isinstance(temp,float):
                 temp = [temp]
             else:
                 temp = [*temp]
@@ -73,7 +73,7 @@ def entrada_test(data: dict, apoio: dict, comprimentoTotal:float):
     for chave_data in data.keys():
         temp = data[chave_data]['pos']
         #Verificar se é esforço pontual
-        if isinstance(temp,int):
+        if isinstance(temp,float):
             for chave_el in elementos.keys():
                 if temp in elementos[chave_el]['Trecho']:
                     contador = 1 + len(elementos[chave_el]['Carregamento'])
